+++ server/src/routes/api/index.ts
@@ -2,7 +2,7 @@ export {};
 
 const express = require("express");
 const router = express.Router();
-const { router: channelRouter } = require("./channels");
+const channelRouter = require("./channels");
 
 router.use("/channels", channelRouter);
